requirements.txt
streamlit
moviepy
whisper
google-generativeai
python-dotenv
opencv-python
pandas
deepface